requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
